+++ mobile/android/android-components/docs/changelog.md
@@ -57,6 +57,9 @@ permalink: /changelog/
 
 * **feature-contextmenu**
   * ⚠️ **This is a breaking change**: `customTabId` is renamed to `tabId`.
+  
+* **browser-menu**
+  * Emit fact on the web extension id when a web extension menu item is clicked.
 
 * **feature-push**
   * ⚠️ **This is a breaking change**: